Splitting one mega keeper file into smaller files
