+++ module/x/gravity/keeper/keeper_batch.go
@@ -1,0 +1,64 @@
+package keeper
+
+import (
+	"github.com/cosmos/cosmos-sdk/store/prefix"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/althea-net/cosmos-gravity-bridge/module/x/gravity/types"
+)
+
+/////////////////////////////
+//      BATCH CONFIRMS     //
+/////////////////////////////
+
+// GetBatchConfirm returns a batch confirmation given its nonce, the token contract, and a validator address
+func (k Keeper) GetBatchConfirm(ctx sdk.Context, nonce uint64, tokenContract string, validator sdk.AccAddress) *types.MsgConfirmBatch {
+	store := ctx.KVStore(k.storeKey)
+	entity := store.Get(types.GetBatchConfirmKey(tokenContract, nonce, validator))
+	if entity == nil {
+		return nil
+	}
+	confirm := types.MsgConfirmBatch{}
+	k.cdc.MustUnmarshalBinaryBare(entity, &confirm)
+	return &confirm
+}
+
+// SetBatchConfirm sets a batch confirmation by a validator
+func (k Keeper) SetBatchConfirm(ctx sdk.Context, batch *types.MsgConfirmBatch) []byte {
+	store := ctx.KVStore(k.storeKey)
+	acc, err := sdk.AccAddressFromBech32(batch.Orchestrator)
+	if err != nil {
+		panic(err)
+	}
+	key := types.GetBatchConfirmKey(batch.TokenContract, batch.Nonce, acc)
+	store.Set(key, k.cdc.MustMarshalBinaryBare(batch))
+	return key
+}
+
+// IterateBatchConfirmByNonceAndTokenContract iterates through all batch confirmations
+// MARK finish-batches: this is where the key is iterated in the old (presumed working) code
+// TODO: specify which nonce this is
+func (k Keeper) IterateBatchConfirmByNonceAndTokenContract(ctx sdk.Context, nonce uint64, tokenContract string, cb func([]byte, types.MsgConfirmBatch) bool) {
+	prefixStore := prefix.NewStore(ctx.KVStore(k.storeKey), types.BatchConfirmKey)
+	prefix := append([]byte(tokenContract), types.UInt64Bytes(nonce)...)
+	iter := prefixStore.Iterator(prefixRange(prefix))
+	defer iter.Close()
+
+	for ; iter.Valid(); iter.Next() {
+		confirm := types.MsgConfirmBatch{}
+		k.cdc.MustUnmarshalBinaryBare(iter.Value(), &confirm)
+		// cb returns true to stop early
+		if cb(iter.Key(), confirm) {
+			break
+		}
+	}
+}
+
+// GetBatchConfirmByNonceAndTokenContract returns the batch confirms
+func (k Keeper) GetBatchConfirmByNonceAndTokenContract(ctx sdk.Context, nonce uint64, tokenContract string) (out []types.MsgConfirmBatch) {
+	k.IterateBatchConfirmByNonceAndTokenContract(ctx, nonce, tokenContract, func(_ []byte, msg types.MsgConfirmBatch) bool {
+		out = append(out, msg)
+		return false
+	})
+	return
+}
