+++ module/x/gravity/keeper/keeper_valset.go
@@ -1,0 +1,254 @@
+package keeper
+
+import (
+	"fmt"
+	"sort"
+	"strconv"
+
+	"github.com/cosmos/cosmos-sdk/store/prefix"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/althea-net/cosmos-gravity-bridge/module/x/gravity/types"
+)
+
+/////////////////////////////
+//     VALSET REQUESTS     //
+/////////////////////////////
+
+// SetValsetRequest returns a new instance of the Gravity BridgeValidatorSet
+// by taking a snapshot of the current set
+// i.e. {"nonce": 1, "memebers": [{"eth_addr": "foo", "power": 11223}]}
+func (k Keeper) SetValsetRequest(ctx sdk.Context) *types.Valset {
+	valset := k.GetCurrentValset(ctx)
+	k.StoreValset(ctx, valset)
+
+	// Store the checkpoint as a legit past valset, this is only for evidence
+	// based slashing. We are storing the checkpoint that will be signed with
+	// the validators Etheruem keys so that we know not to slash them if someone
+	// attempts to submit the signature of this validator set as evidence of bad behavior
+	checkpoint := valset.GetCheckpoint(k.GetGravityID(ctx))
+	k.SetPastEthSignatureCheckpoint(ctx, checkpoint)
+
+	ctx.EventManager().EmitEvent(
+		sdk.NewEvent(
+			types.EventTypeMultisigUpdateRequest,
+			sdk.NewAttribute(sdk.AttributeKeyModule, types.ModuleName),
+			sdk.NewAttribute(types.AttributeKeyContract, k.GetBridgeContractAddress(ctx)),
+			sdk.NewAttribute(types.AttributeKeyBridgeChainID, strconv.Itoa(int(k.GetBridgeChainID(ctx)))),
+			sdk.NewAttribute(types.AttributeKeyMultisigID, fmt.Sprint(valset.Nonce)),
+			sdk.NewAttribute(types.AttributeKeyNonce, fmt.Sprint(valset.Nonce)),
+		),
+	)
+
+	return valset
+}
+
+// StoreValset is for storing a valiator set at a given height
+func (k Keeper) StoreValset(ctx sdk.Context, valset *types.Valset) {
+	store := ctx.KVStore(k.storeKey)
+	valset.Height = uint64(ctx.BlockHeight())
+	store.Set(types.GetValsetKey(valset.Nonce), k.cdc.MustMarshalBinaryBare(valset))
+	k.SetLatestValsetNonce(ctx, valset.Nonce)
+}
+
+//  SetLatestValsetNonce sets the latest valset nonce
+func (k Keeper) SetLatestValsetNonce(ctx sdk.Context, nonce uint64) {
+	store := ctx.KVStore(k.storeKey)
+	store.Set(types.LatestValsetNonce, types.UInt64Bytes(nonce))
+}
+
+// StoreValsetUnsafe is for storing a valiator set at a given height
+func (k Keeper) StoreValsetUnsafe(ctx sdk.Context, valset *types.Valset) {
+	store := ctx.KVStore(k.storeKey)
+	store.Set(types.GetValsetKey(valset.Nonce), k.cdc.MustMarshalBinaryBare(valset))
+	k.SetLatestValsetNonce(ctx, valset.Nonce)
+}
+
+// HasValsetRequest returns true if a valset defined by a nonce exists
+func (k Keeper) HasValsetRequest(ctx sdk.Context, nonce uint64) bool {
+	store := ctx.KVStore(k.storeKey)
+	return store.Has(types.GetValsetKey(nonce))
+}
+
+// DeleteValset deletes the valset at a given nonce from state
+func (k Keeper) DeleteValset(ctx sdk.Context, nonce uint64) {
+	ctx.KVStore(k.storeKey).Delete(types.GetValsetKey(nonce))
+}
+
+// GetLatestValsetNonce returns the latest valset nonce
+func (k Keeper) GetLatestValsetNonce(ctx sdk.Context) uint64 {
+	store := ctx.KVStore(k.storeKey)
+	bytes := store.Get(types.LatestValsetNonce)
+
+	if len(bytes) == 0 {
+		return 0
+	}
+	return types.UInt64FromBytes(bytes)
+}
+
+// GetValset returns a valset by nonce
+func (k Keeper) GetValset(ctx sdk.Context, nonce uint64) *types.Valset {
+	store := ctx.KVStore(k.storeKey)
+	bz := store.Get(types.GetValsetKey(nonce))
+	if bz == nil {
+		return nil
+	}
+	var valset types.Valset
+	k.cdc.MustUnmarshalBinaryBare(bz, &valset)
+	return &valset
+}
+
+// IterateValsets retruns all valsetRequests
+func (k Keeper) IterateValsets(ctx sdk.Context, cb func(key []byte, val *types.Valset) bool) {
+	prefixStore := prefix.NewStore(ctx.KVStore(k.storeKey), types.ValsetRequestKey)
+	iter := prefixStore.ReverseIterator(nil, nil)
+	defer iter.Close()
+	for ; iter.Valid(); iter.Next() {
+		var valset types.Valset
+		k.cdc.MustUnmarshalBinaryBare(iter.Value(), &valset)
+		// cb returns true to stop early
+		if cb(iter.Key(), &valset) {
+			break
+		}
+	}
+}
+
+// GetValsets returns all the validator sets in state
+func (k Keeper) GetValsets(ctx sdk.Context) (out []*types.Valset) {
+	k.IterateValsets(ctx, func(_ []byte, val *types.Valset) bool {
+		out = append(out, val)
+		return false
+	})
+	sort.Sort(types.Valsets(out))
+	return
+}
+
+// GetLatestValset returns the latest validator set in state
+func (k Keeper) GetLatestValset(ctx sdk.Context) (out *types.Valset) {
+	latestValsetNonce := k.GetLatestValsetNonce(ctx)
+	out = k.GetValset(ctx, latestValsetNonce)
+	return
+}
+
+// setLastSlashedValsetNonce sets the latest slashed valset nonce
+func (k Keeper) SetLastSlashedValsetNonce(ctx sdk.Context, nonce uint64) {
+	store := ctx.KVStore(k.storeKey)
+	store.Set(types.LastSlashedValsetNonce, types.UInt64Bytes(nonce))
+}
+
+// GetLastSlashedValsetNonce returns the latest slashed valset nonce
+func (k Keeper) GetLastSlashedValsetNonce(ctx sdk.Context) uint64 {
+	store := ctx.KVStore(k.storeKey)
+	bytes := store.Get(types.LastSlashedValsetNonce)
+
+	if len(bytes) == 0 {
+		return 0
+	}
+	return types.UInt64FromBytes(bytes)
+}
+
+// SetLastUnBondingBlockHeight sets the last unbonding block height
+func (k Keeper) SetLastUnBondingBlockHeight(ctx sdk.Context, unbondingBlockHeight uint64) {
+	store := ctx.KVStore(k.storeKey)
+	store.Set(types.LastUnBondingBlockHeight, types.UInt64Bytes(unbondingBlockHeight))
+}
+
+// GetLastUnBondingBlockHeight returns the last unbonding block height
+func (k Keeper) GetLastUnBondingBlockHeight(ctx sdk.Context) uint64 {
+	store := ctx.KVStore(k.storeKey)
+	bytes := store.Get(types.LastUnBondingBlockHeight)
+
+	if len(bytes) == 0 {
+		return 0
+	}
+	return types.UInt64FromBytes(bytes)
+}
+
+// GetUnSlashedValsets returns all the unslashed validator sets in state
+func (k Keeper) GetUnSlashedValsets(ctx sdk.Context, maxHeight uint64) (out []*types.Valset) {
+	lastSlashedValsetNonce := k.GetLastSlashedValsetNonce(ctx)
+	k.IterateValsetBySlashedValsetNonce(ctx, lastSlashedValsetNonce, maxHeight, func(_ []byte, valset *types.Valset) bool {
+		if valset.Nonce > lastSlashedValsetNonce {
+			out = append(out, valset)
+		}
+		return false
+	})
+	return
+}
+
+// IterateValsetBySlashedValsetNonce iterates through all valset by last slashed valset nonce in ASC order
+func (k Keeper) IterateValsetBySlashedValsetNonce(ctx sdk.Context, lastSlashedValsetNonce uint64, maxHeight uint64, cb func([]byte, *types.Valset) bool) {
+	prefixStore := prefix.NewStore(ctx.KVStore(k.storeKey), types.ValsetRequestKey)
+	iter := prefixStore.Iterator(types.UInt64Bytes(lastSlashedValsetNonce), types.UInt64Bytes(maxHeight))
+	defer iter.Close()
+
+	for ; iter.Valid(); iter.Next() {
+		var valset types.Valset
+		k.cdc.MustUnmarshalBinaryBare(iter.Value(), &valset)
+		// cb returns true to stop early
+		if cb(iter.Key(), &valset) {
+			break
+		}
+	}
+}
+
+/////////////////////////////
+//     VALSET CONFIRMS     //
+/////////////////////////////
+
+// GetValsetConfirm returns a valset confirmation by a nonce and validator address
+func (k Keeper) GetValsetConfirm(ctx sdk.Context, nonce uint64, validator sdk.AccAddress) *types.MsgValsetConfirm {
+	store := ctx.KVStore(k.storeKey)
+	entity := store.Get(types.GetValsetConfirmKey(nonce, validator))
+	if entity == nil {
+		return nil
+	}
+	confirm := types.MsgValsetConfirm{}
+	k.cdc.MustUnmarshalBinaryBare(entity, &confirm)
+	return &confirm
+}
+
+// SetValsetConfirm sets a valset confirmation
+func (k Keeper) SetValsetConfirm(ctx sdk.Context, valsetConf types.MsgValsetConfirm) []byte {
+	store := ctx.KVStore(k.storeKey)
+	addr, err := sdk.AccAddressFromBech32(valsetConf.Orchestrator)
+	if err != nil {
+		panic(err)
+	}
+	key := types.GetValsetConfirmKey(valsetConf.Nonce, addr)
+	store.Set(key, k.cdc.MustMarshalBinaryBare(&valsetConf))
+	return key
+}
+
+// GetValsetConfirms returns all validator set confirmations by nonce
+func (k Keeper) GetValsetConfirms(ctx sdk.Context, nonce uint64) (confirms []*types.MsgValsetConfirm) {
+	prefixStore := prefix.NewStore(ctx.KVStore(k.storeKey), types.ValsetConfirmKey)
+	start, end := prefixRange(types.UInt64Bytes(nonce))
+	iterator := prefixStore.Iterator(start, end)
+
+	defer iterator.Close()
+
+	for ; iterator.Valid(); iterator.Next() {
+		confirm := types.MsgValsetConfirm{}
+		k.cdc.MustUnmarshalBinaryBare(iterator.Value(), &confirm)
+		confirms = append(confirms, &confirm)
+	}
+
+	return confirms
+}
+
+// IterateValsetConfirmByNonce iterates through all valset confirms by validator set nonce in ASC order
+func (k Keeper) IterateValsetConfirmByNonce(ctx sdk.Context, nonce uint64, cb func([]byte, types.MsgValsetConfirm) bool) {
+	prefixStore := prefix.NewStore(ctx.KVStore(k.storeKey), types.ValsetConfirmKey)
+	iter := prefixStore.Iterator(prefixRange(types.UInt64Bytes(nonce)))
+	defer iter.Close()
+
+	for ; iter.Valid(); iter.Next() {
+		confirm := types.MsgValsetConfirm{}
+		k.cdc.MustUnmarshalBinaryBare(iter.Value(), &confirm)
+		// cb returns true to stop early
+		if cb(iter.Key(), confirm) {
+			break
+		}
+	}
+}
